+++ src/macro.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
+import { basename, extname } from 'path';
 import { Lazy } from './common/lazy';
 import { MacroOptions, parseOptions } from './macroOptions';
-import { basename } from 'path';
 
 export type MacroId = string;
 
@@ -31,6 +31,7 @@ export class Macro {
   }
 
   public get shortName(): string {
-    return basename(this.uri.fsPath);
+    const path = this.uri.fsPath;
+    return basename(path, extname(path));
   }
 }