+++ src/macrosApi.ts
@@ -1,13 +1,19 @@
 import * as vscode from 'vscode';
 
+export interface MacroApi {
+  /**
+   * Id of current macro execution.
+   */
+  readonly runId: string;
+  /**
+   * URI of current macro. It can be undefined if running from an in-memory buffer.
+   */
+  readonly uri: vscode.Uri | undefined;
+}
+
 export interface MacrosApi {
   /**
    * Current macro
    */
-  macro: {
-    /**
-     * URI of current macro. It can be undefined if running from an in-memory buffer.
-     */
-    readonly uri: vscode.Uri | undefined;
-  }
+  macro: MacroApi
 }